requirements.txt
httpx
python-telegram-bot
python-dotenv
beautifulsoup4
flask
asgiref
pandas
pyrogram
TgCrypto
